Fixing bugs in Reconstruction code. Updating user interface with some tweaks. Fixing bugs with the file paths not being completely correct for each platform. Fixing compile warnings in AngFileHelper with comparisons between int and size_t. Adding a check after the thread is finished to rescan the output directory for the output files.

+++ Source/AIM/ANG/AngFileHelper.cpp
@@ -37,7 +37,7 @@ void AngFileHelper::loadData(Voxel voxels[],
   const double badeuler = 12.566370010375977;
   size_t index = 0;
   int nSlices = m_ZIndexEnd - m_ZIndexStart;
-  for (size_t k = 0; k < nSlices; k++)
+  for (int k = 0; k < nSlices; k++)
   {
     if (m_Cancel == true) { break; }
     {
